// client/src/store/constants.tsx
type Obj = Record<string, string>;
const constants: Obj = {
  LOGIN_SUCCESS: 'LOGIN_SUCCESS',
  LOGIN_FAILED: 'LOGIN_FAILED',
  LOGIN_LOADING: 'LOGIN_LOADING',

  REGISTRATION_SUCCESS: 'REGISTRATION_SUCCESS',
  REGISTRATION_FAILED: 'REGISTRATION_FAILED',
  REGISTRATION_LOADING: 'REGISTRATION_LOADING',

  LOGOUT_SUCCESS: 'LOGOUT_SUCCESS',
  LOGOUT_LOADING: 'LOGOUT_LOADING',
  LOGOUT_FAILED: 'LOGOUT_FAILED',

  GET_ROLE_SUCCESS: 'GET_ROLE_SUCCESS',
  GET_ROLE_LOADING: 'GET_ROLE_LOADING',
  GET_ROLE_FAILED: 'GET_ROLE_FAILED',

  GET_TYPES_SUCCESS: 'GET_TYPES_SUCCESS',
  GET_TYPES_LOADING: 'GET_TYPES_LOADING',
  GET_TYPES_FAILED: 'GET_TYPES_FAILED',

  ADD_TYPE_SUCCESS: 'ADD_TYPE_SUCCESS',
  ADD_TYPE_LOADING: 'ADD_TYPE_LOADING',
  ADD_TYPE_FAILED: 'ADD_TYPE_FAILED',

  EDIT_TYPE_SUCCESS: 'EDIT_TYPE_SUCCESS',
  EDIT_TYPE_LOADING: 'EDIT_TYPE_LOADING',
  EDIT_TYPE_FAILED: 'EDIT_TYPE_FAILED',

  DELETE_TYPE_SUCCESS: 'DELETE_TYPE_SUCCESS',
  DELETE_TYPE_LOADING: 'DELETE_TYPE_LOADING',
  DELETE_TYPE_FAILED: 'DELETE_TYPE_FAILED',

  GET_BRANDS_SUCCESS: 'GET_BRANDS_SUCCESS',
  GET_BRANDS_LOADING: 'GET_BRANDS_LOADING',
  GET_BRANDS_FAILED: 'GET_BRANDS_FAILED',

  ADD_BRAND_SUCCESS: 'ADD_BRAND_SUCCESS',
  ADD_BRAND_LOADING: 'ADD_BRAND_LOADING',
  ADD_BRAND_FAILED: 'ADD_BRAND_FAILED',

  EDIT_BRAND_SUCCESS: 'EDIT_BRAND_SUCCESS',
  EDIT_BRAND_LOADING: 'EDIT_BRAND_LOADING',
  EDIT_BRAND_FAILED: 'EDIT_BRAND_FAILED',

  DELETE_BRAND_SUCCESS: 'DELETE_BRAND_SUCCESS',
  DELETE_BRAND_LOADING: 'DELETE_BRAND_LOADING',
  DELETE_BRAND_FAILED: 'DELETE_BRAND_FAILED',

  GET_DRINKS_SUCCESS: 'GET_DRINKS_SUCCESS',
  GET_DRINKS_LOADING: 'GET_DRINKS_LOADING',
  GET_DRINKS_FAILED: 'GET_DRINKS_FAILED',

  RECEIPT_DRINK_SUCCESS: 'RECEIPT_DRINK_SUCCESS',
  RECEIPT_DRINK_LOADING: 'RECEIPT_DRINK_LOADING',
  RECEIPT_DRINK_FAILED: 'RECEIPT_DRINK_FAILED',

  GET_DRINK_SUCCESS: 'GET_DRINK_SUCCESS',
  GET_DRINK_LOADING: 'GET_DRINK_LOADING',
  GET_DRINK_FAILED: 'GET_DRINK_FAILED',

  GET_DRINKS_PAGINATED_SUCCESS: 'GET_DRINKS_PAGINATED_SUCCESS',
  GET_DRINKS_PAGINATED_LOADING: 'GET_DRINKS_PAGINATED_LOADING',
  GET_DRINKS_PAGINATED_FAILED: 'GET_DRINKS_PAGINATED_FAILED',

  ADD_DRINK_SUCCESS: 'ADD_DRINK_SUCCESS',
  ADD_DRINK_LOADING: 'ADD_DRINK_LOADING',
  ADD_DRINK_FAILED: 'ADD_DRINK_FAILED',

  EDIT_DRINK_SUCCESS: 'EDIT_DRINK_SUCCESS',
  EDIT_DRINK_LOADING: 'EDIT_DRINK_LOADING',
  EDIT_DRINK_FAILED: 'EDIT_DRINK_FAILED',

  DELETE_DRINK_SUCCESS: 'DELETE_DRINK_SUCCESS',
  DELETE_DRINK_LOADING: 'DELETE_DRINK_LOADING',
  DELETE_DRINK_FAILED: 'DELETE_DRINK_FAILED',

  GET_ALL_COUPONS_SUCCESS: 'GET_ALL_COUPONS_SUCCESS',
  GET_ALL_COUPONS_LOADING: 'GET_ALL_COUPONS_LOADING',
  GET_ALL_COUPONS_FAILED: 'GET_ALL_COUPONS_FAILED',

  GET_MY_COUPONS_SUCCESS: 'GET_MY_COUPONS_SUCCESS',
  GET_MY_COUPONS_LOADING: 'GET_MY_COUPONS_LOADING',
  GET_MY_COUPONS_FAILED: 'GET_MY_COUPONS_FAILED',

  ADD_COUPON_SUCCESS: 'ADD_COUPON_SUCCESS',
  ADD_COUPON_LOADING: 'ADD_COUPON_LOADING',
  ADD_COUPON_FAILED: 'ADD_COUPON_FAILED',

  ADD_COUPON_TO_ME_SUCCESS: 'ADD_COUPON_TO_ME_SUCCESS',
  ADD_COUPON_TO_ME_LOADING: 'ADD_COUPON_TO_ME_LOADING',
  ADD_COUPON_TO_ME_FAILED: 'ADD_COUPON_TO_ME_FAILED',

  DELETE_COUPON_SUCCESS: 'DELETE_COUPON_SUCCESS',
  DELETE_COUPON_LOADING: 'DELETE_COUPON_LOADING',
  DELETE_COUPON_FAILED: 'DELETE_COUPON_FAILED',

  GET_DRINKS_BASKET_SUCCESS: 'GET_DRINKS_BASKET_SUCCESS',
  GET_DRINKS_BASKET_LOADING: 'GET_DRINKS_BASKET_LOADING',
  GET_DRINKS_BASKET_FAILED: 'GET_DRINKS_BASKET_FAILED',

  ADD_DRINK_TO_ME_SUCCESS: 'ADD_DRINK_TO_ME_SUCCESS',
  ADD_DRINK_TO_ME_LOADING: 'ADD_DRINK_TO_ME_LOADING',
  ADD_DRINK_TO_ME_FAILED: 'ADD_DRINK_TO_ME_FAILED',

  DELETE_DRINK_BASKET_SUCCESS: 'DELETE_DRINK_BASKET_SUCCESS',
  DELETE_DRINK_BASKET_LOADING: 'DELETE_DRINK_BASKET_LOADING',
  DELETE_DRINK_BASKET_FAILED: 'DELETE_DRINK_BASKET_FAILED',

  ADD_ORDER_SUCCESS: 'ADD_ORDER_SUCCESS',
  ADD_ORDER_LOADING: 'ADD_ORDER_LOADING',
  ADD_ORDER_FAILED: 'ADD_ORDER_FAILED',

  GET_USER_ORDERS_SUCCESS: 'GET_USER_ORDERS_SUCCESS',
  GET_USER_ORDERS_LOADING: 'GET_USER_ORDERS_LOADING',
  GET_USER_ORDERS_FAILED: 'GET_USER_ORDERS_FAILED',

  GET_ORDERS_SUCCESS: 'GET_ORDERS_SUCCESS',
  GET_ORDERS_LOADING: 'GET_ORDERS_LOADING',
  GET_ORDERS_FAILED: 'GET_ORDERS_FAILED',

  EXECUTE_ORDER_SUCCESS: 'EXECUTE_ORDER_SUCCESS',
  EXECUTE_ORDER_LOADING: 'EXECUTE_ORDER_LOADING',
  EXECUTE_ORDER_FAILED: 'EXECUTE_ORDER_FAILED',

  GET_REPORT_SUCCESS: 'GET_REPORT_SUCCESS',
  GET_REPORT_LOADING: 'GET_REPORT_LOADING',
  GET_REPORT_FAILED: 'GET_REPORT_FAILED',
};